Small fixes

+++ Common/src/BitmapLoader.cpp
@@ -79,9 +79,9 @@ std::unique_ptr<Image> BitmapLoader::LoadBitmapImage()
 	// Load data
 	file.seekg(offset); // go to start of data
 
-	for (uint32_t y = 0; y < bmpInfo.biHeight; y++)
+	for (uint32_t y = 0; y < static_cast<uint32_t>(bmpInfo.biHeight); y++)
 	{
-		for (uint32_t x = 0; x < bmpInfo.biWidth; x++)
+		for (uint32_t x = 0; x < static_cast<uint32_t>(bmpInfo.biWidth); x++)
 		{
 			// read pixel
 			uint8_t data[3];
